✨️#76: 상품 응답 DTO 통일
개발 편의 상 한 개의 DTO 응답으로 통일합니다.
